+++ test/package.test.js
@@ -11,8 +11,7 @@ describe('@authnomicon/prompts', function() {
     it('should have assembly metadata', function() {
       expect(json.assembly.namespace).to.equal('org.authnomicon/prompts');
       expect(json.assembly.components).to.deep.equal([
-        'http/registry',
-        'http/logout'
+        'http/registry'
       ]);
     });
   });